feat:deleting profile with saga

+++ Nistagram/activity-service/src/main/java/app/activityservice/service/ProfileServiceImpl.java
@@ -75,9 +75,11 @@ public class ProfileServiceImpl implements ProfileService {
 	}
 
 	@Override
+	@Transactional
 	public void deleteProfileByUsername(String username) {
 		Profile profile = profileRepository.findByUsername(username);
-		profileRepository.delete(profile);
+		if(profile != null)
+			profileRepository.delete(profile);
 		
 	}
